# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inova Banco API",
        "description": "Criado por Leo França Dev",
        "version": "1.0"
    },
    "paths": {
        "/user": {
            "post": {
                "tags": [
                    "Usuários"
                ],
                "operationId": "fdaf6e20ea63c9e43bacb0c1f25012d5",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            },
                                            "user_type_id": {
                                                "type": "integer"
                                            },
                                            "balance": {
                                                "type": "string"
                                            },
                                            "cpf_cnpj": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "leofranca",
                                    "email": "leo@email.com",
                                    "password": "123456"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "leofranca"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "leo@email.com"
                                        },
                                        "user_type_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "balance": {
                                            "type": "string",
                                            "example": "100.00"
                                        },
                                        "cpf_cnpj": {
                                            "type": "string",
                                            "example": "12345678901"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id}": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "operationId": "1ec60d4ee968175164a94c21f897f8ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Leo"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "leo@email.com"
                                        },
                                        "user_type_id": {
                                            "type": "integer",
                                            "example": 58
                                        },
                                        "balance": {
                                            "type": "string",
                                            "example": "10.2"
                                        },
                                        "cpf_cnpj": {
                                            "type": "string",
                                            "example": "57649769000152"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/transfer": {
            "post": {
                "tags": [
                    "Usuários"
                ],
                "operationId": "2beda4ceaeaaffd5cb0ee38da312b179",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "value": {
                                                "type": "string"
                                            },
                                            "payer": {
                                                "type": "integer"
                                            },
                                            "payee": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "value": "100.00",
                                    "payer": 2,
                                    "payee": 1
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transferência realizada com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "falhou"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Usuários",
            "description": "Usuários"
        }
    ]
}